refactor: rename `CsrRenderedInitials` to `CsrInitialsRenderer`
