+++ src/js/visual/index.js
@@ -1,6 +1,6 @@
 const csrAnimation = require("./csr/animation-base");
 const csrChangeFrameAnimation = require("./csr/animation-change-frame");
-const csrInitialsRenderer = require("./csr/initials-renderer");
+const csrRenderedInitials = require("./csr/rendered-initials");
 const csrTextureRenderer = require("./csr/texture-renderer");
 const csrUtils = require("./csr/utils");
 const configuratorCsr = require("./configurator-csr");
@@ -10,7 +10,7 @@ const visual = require("./visual");
 
 Object.assign(module.exports, csrAnimation);
 Object.assign(module.exports, csrChangeFrameAnimation);
-Object.assign(module.exports, csrInitialsRenderer);
+Object.assign(module.exports, csrRenderedInitials);
 Object.assign(module.exports, csrTextureRenderer);
 Object.assign(module.exports, csrUtils);
 Object.assign(module.exports, configuratorCsr);
